feat(bridge-protocol, registry-protocol): add message types and constants for communication
