+++ packages/bridge-protocol/src/model.ts
@@ -1,0 +1,33 @@
+import type { APP_MESSAGE_TYPE, HOST_MESSAGE_TYPE } from './constant';
+
+export interface ContentBbox {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
+
+export interface LoadFilePayload {
+  data: unknown;
+  fileName: string;
+}
+
+export type HostMessage =
+  | {
+      type: typeof HOST_MESSAGE_TYPE.LOAD;
+      payload: { content: string; fileName: string };
+    }
+  | { type: typeof HOST_MESSAGE_TYPE.SAVED }
+  | { type: typeof HOST_MESSAGE_TYPE.LOAD_FILE; payload: LoadFilePayload };
+
+export type AppMessage =
+  | { type: typeof APP_MESSAGE_TYPE.READY }
+  | { type: typeof APP_MESSAGE_TYPE.WEBVIEW_READY }
+  | { type: typeof APP_MESSAGE_TYPE.SAVE; payload: { content: string } }
+  | {
+      type: typeof APP_MESSAGE_TYPE.RENDER_COMPLETE;
+      payload?: ContentBbox;
+    };
+
+export type PayloadOf<U extends { type: string }, T extends U['type']> =
+  Extract<U, { type: T }> extends { payload: infer P } ? P : undefined;
