+++ packages/bridge-protocol/tsconfig.json
@@ -1,10 +1,4 @@
 {
-  "extends": "@lemoncode/typescript-config/base",
-  "include": ["src"],
-  "compilerOptions": {
-    "target": "ES2024",
-    "lib": ["ES2024"],
-    "noEmit": true,
-    "rootDir": "src"
-  }
+  "extends": "@lemoncode/typescript-config/node",
+  "include": ["src"]
 }
