+++ packages/registry-protocol/src/constant.ts
@@ -2,3 +2,4 @@ export const TOKEN_HEADER = 'x-quickmock-token';
 export const LOOPBACK_HOST = '127.0.0.1';
 export const DOCUMENT_ROUTE = '/document';
 export const PORT_TOKEN_SEPARATOR = ' ';
+export const WORKSPACE_HASH_LENGTH = 8;
