+++ packages/registry-protocol/src/index.ts
@@ -1,2 +1,2 @@
-export * from './consts';
+export * from './constant';
 export * from './utils';
