+++ packages/registry-protocol/tsconfig.json
@@ -1,11 +1,4 @@
 {
-  "extends": "@lemoncode/typescript-config/base",
-  "include": ["src"],
-  "compilerOptions": {
-    "target": "ES2024",
-    "types": ["node"],
-    "lib": ["ES2024"],
-    "noEmit": true,
-    "rootDir": "src"
-  }
+  "extends": "@lemoncode/typescript-config/node",
+  "include": ["src"]
 }
